services/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
neo4j==5.15.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.0
pandas==2.1.4
numpy==1.24.3
loguru==0.7.2
httpx==0.25.2
